pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gtfs_upcoming"
dependencies = [
    "click==8.1.3",
    "gtfs-realtime-bindings==1.0.0",
    "prometheus-client==0.17.0",
    "protobuf<6.0,>=5.0",
    "opentelemetry-distro==0.55b1",
    "opentelemetry-exporter-otlp==1.34.1",
    "opentelemetry-instrumentation-asyncio==0.55b1",
    "opentelemetry-instrumentation-dbapi==0.55b1",
    "opentelemetry-instrumentation-logging==0.55b1",
    "opentelemetry-instrumentation-sqlite3==0.55b1",
    "opentelemetry-instrumentation-threading==0.55b1",
    "opentelemetry-instrumentation-urllib==0.55b1",
    "opentelemetry-instrumentation-wsgi==0.55b1",
    "opentelemetry-instrumentation-click==0.55b1",
    "opentelemetry-instrumentation-grpc==0.55b1",
    "opentelemetry-instrumentation-requests==0.55b1",
    "opentelemetry-instrumentation-urllib3==0.55b1",
]
requires-python = ">=3.9"
authors = [
  {name = "Sean Rees", email = "sean@erifax.org"}
]
maintainers = [
  {name = "Sean Rees", email = "sean@erifax.org"}
]
description = "Easy-to-consume realtime GTFS feed"
readme = "README.md"
license = "MIT"
license-files = ["LICEN[CS]E.*"]
keywords = ["gtfs", "realtime", "vicroads", "nta"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/seanrees/gtfs-upcoming"
Documentation = "https://github.com/seanrees/gtfs-upcoming"
Repository = "https://github.com/seanrees/gtfs-upcoming"
"Bug Tracker" = "https://github.com/seanrees/gtfs-upcoming/issues"
Changelog = "https://github.com/seanrees/gtfs-upcoming/blob/main/CHANGELOG.md"

[project.scripts]
gtfs-upcoming = "gtfs_upcoming.main:main"

[tool]

[tool.hatch.version]
path = "src/gtfs_upcoming/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/gtfs_upcoming tests}"

[tool.coverage.run]
source_pkgs = ["gtfs_upcoming", "tests"]
branch = true
parallel = true
omit = [
  "src/gtfs_upcoming/__about__.py",
]

[tool.coverage.paths]
gtfs_upcoming = ["src/gtfs_upcoming", "*/gtfs-upcoming/src/gtfs_upcoming"]
tests = ["tests", "*/gtfs-upcoming/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
[tool.hatch.envs.test]
dependencies = [
  "pytest"
]

[tool.ruff.lint]
# DTZ001,5,7: we want to use naive datetimes because the GTFS schedule is in local time.
# N802,3,6,N815: "function name should be lowercase; perhaps a change for later as it's a lot of work
# TCH003: wants to only load if doing type-checking, this seems silly
# TID252: relative imports are nice
ignore = ["DTZ001", "DTZ005", "DTZ006", "DTZ007", "TCH003", "TID252"]

[tool.ruff.lint.extend-per-file-ignores]
# SLF001: tests can access internal state
"tests/*" = ["SLF001"]